+++ cli/lib/command/internal_command_about.dart
@@ -2,22 +2,31 @@ import 'package:stax/command/internal_command.dart';
 import 'package:stax/context/context.dart';
 
 class InternalCommandAbout extends InternalCommand {
-  InternalCommandAbout()
-    : super('about', 'Shows information about the stax.');
+  InternalCommandAbout() : super('about', 'Shows information about the stax.');
 
   @override
   void run(final List<String> args, Context context) {
-    context.printToConsole('stax - A Git workflow tool');
-    context.printToConsole('');
-    context.printToConsole(
-      'stax is a command-line tool that helps you manage your Git workflow more efficiently.',
-    );
-    context.printToConsole(
-      'It provides a set of commands to simplify common Git operations and maintain a clean repository structure.',
-    );
-    context.printToConsole('');
-    context.printToConsole(
-      'For more information, visit: https://staxforgit.com/',
-    );
+    context.printToConsole('''
+stax - manage git branches and stack PRs
+
+For more information, visit: https://staxforgit.com/
+
+Copyright (C) ${DateTime.now().year}  Taras Mazepa
+
+This program is free software: you can redistribute it and/or modify
+it under the terms of the GNU General Public License as published by
+the Free Software Foundation, either version 3 of the License, or
+(at your option) any later version.
+
+This program is distributed in the hope that it will be useful,
+but WITHOUT ANY WARRANTY; without even the implied warranty of
+MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+GNU General Public License for more details.
+
+You should have received a copy of the GNU General Public License
+along with this program.  If not, see <https://www.gnu.org/licenses/>.
+
+https://github.com/TarasMazepa/stax/blob/main/LICENSE
+'''.trim());
   }
 }