+++ cli/test/cli/about_test.dart
@@ -6,16 +6,28 @@ import 'base/cli_group.dart';
 void main() {
   cliGroup('about', (setup) {
     test('about', () {
-      expect(
-        setup.runLiveStaxSync(['about']).stdout,
-        '''stax - A Git workflow tool
-
-stax is a command-line tool that helps you manage your Git workflow more efficiently.
-It provides a set of commands to simplify common Git operations and maintain a clean repository structure.
+      expect(setup.runLiveStaxSync(['about']).stdout, '''
+stax - manage git branches and stack PRs
 
 For more information, visit: https://staxforgit.com/
-''',
-      );
+
+Copyright (C) 2025  Taras Mazepa
+
+This program is free software: you can redistribute it and/or modify
+it under the terms of the GNU General Public License as published by
+the Free Software Foundation, either version 3 of the License, or
+(at your option) any later version.
+
+This program is distributed in the hope that it will be useful,
+but WITHOUT ANY WARRANTY; without even the implied warranty of
+MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+GNU General Public License for more details.
+
+You should have received a copy of the GNU General Public License
+along with this program.  If not, see <https://www.gnu.org/licenses/>.
+
+https://github.com/TarasMazepa/stax/blob/main/LICENSE
+''');
     });
   });
 }